pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-classifier-template"
version = "0.1.0"
description = "Local-first, customizable classifier template for production-minded Python projects"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
  { name = "zaber-dev" }
]
dependencies = [
  "PyYAML>=6.0"
]

[project.optional-dependencies]
ml = [
  "scikit-learn>=1.4",
  "joblib>=1.4"
]
dev = [
  "pytest>=8.2",
  "ruff>=0.6.0",
  "mypy>=1.10",
  "build>=1.2.0"
]

[project.scripts]
ai-classifier = "ai_classifier.cli.commands:main"

[project.urls]
Homepage = "https://github.com/zaber-dev/ai-classifier-template"
Repository = "https://github.com/zaber-dev/ai-classifier-template"
Issues = "https://github.com/zaber-dev/ai-classifier-template/issues"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
